Move "full_configuration" fixture to tools/pytest_fixtures/webdriver.py
Issue: #59157
