+++ core-aam/aamtests/conftest.py
@@ -6,6 +6,7 @@ webdriver_tests_path = Path(__file__).parent.parent.parent / "webdriver"
 sys.path.insert(0, str(webdriver_tests_path))
 
 pytest_plugins = (
+    "pytest_fixtures.webdriver",
     "tests.support.fixtures",
     "tests.support.classic.fixtures",
     "aamtests.support.fixtures_a11y_api",
