+++ webdriver/tests/conftest.py
@@ -1,4 +1,5 @@
 pytest_plugins = (
+    "pytest_fixtures.webdriver",
     "tests.support.fixtures",
     "tests.support.bidi.fixtures",
     "tests.support.classic.fixtures",
